feat(handdecomp): port StartEventEarthQuake (sub_8012CEC) → matching C
screen NEAR: s8 play_sound sign-ext (lsr->asr); int param + (s8) cast.
PlaySoundEffect macro inlines to JP gPlaySt-check+m4aSongNumStart. mc 7641→7642.

+++ include/bmfx.h
@@ -99,7 +99,7 @@ void ShinningStonefx_Delay(struct ProcShinningStonefx * proc);
 // ??? NewBoomFogFx(???);
 // ??? CheckBoomFogFxExists(???);
 void EventEarthQuakeMain(struct ProcBmFx * proc);
-void StartEventEarthQuake(u8 type, u8 direction, s8 play_sound);
+void StartEventEarthQuake(u8 type, u8 direction, int play_sound);
 void EndEventEarthQuake(void);
 // ??? StoneShatterEvent_OnEnd(???);
 void StartStoneShatterAnim(struct Unit *, ProcPtr);
